# search: add index batching