Increase indexing throughput by batching events per index.

This PR introduces a per index batch processor. Everytime a new event is ingested:
- the event is dispatched to a queue, one queue per index.
- Within each queue, events are indexed in small batch. The Write/Delete methods have been replaced with a single BatchIndex operation.
- Processor are concurrent with one-another: events in different indices can be processed concurrently

Add a few integration tests for indexing:
- A couple simple tests to verify the implementation for the SearchBackend. This really aims to be a starting point to verify the implementation is correct
- Add a performance test for the Search Backend and the Search server


## Benchmark

Here are 3 run of TestIntegrationBenchmarkResourceServer showing before/after. As you can see, throughput went from ~35 events/s to ~2k per index (x65)
```
## Before with 100 resources
=== RUN   TestIntegrationBenchmarkResourceServer
    Benchmark Configuration: Workers=1, Resources=100, Namespaces=1, Groups=1, Resource Types=1
    Storage Benchmark Results:
    Total Duration: 39.488875ms
    
    Storage Write Count: 100
    Storage Write Throughput: 2532.36 writes/sec
    P50 Write Latency: 326.125µs
    P90 Write Latency: 349.542µs
    P99 Write Latency: 6.472458ms
    
    Index Latency Results:
    Indexing Throughput: 35.48 events/sec
    P50 Index Latency: 1.473s
    P90 Index Latency: 2.573s
    P99 Index Latency: 2.779s

## After with 100 resources
=== RUN   TestIntegrationBenchmarkResourceServer
    Benchmark Configuration: Workers=1, Resources=100, Namespaces=1, Groups=1, Resource Types=1
        
    Storage Benchmark Results:
    Total Duration: 44.591334ms
    Storage Write Count: 100
    Storage Write Throughput: 2242.59 writes/sec
    P50 Write Latency: 322.25µs
    P90 Write Latency: 399.709µs
    P99 Write Latency: 9.143542ms
    
    Index Latency Results:
    Indexing Throughput: 1163.41 events/sec
    P50 Index Latency: 0.041s
    P90 Index Latency: 0.053s
    P99 Index Latency: 0.059s

## After with 10k resources  
=== RUN   TestIntegrationBenchmarkResourceServer
 Benchmark Configuration: Workers=1, Resources=10000, Namespaces=1, Groups=1, Resource Types=1

 Storage Benchmark Results:
 Total Duration: 4.628090541s
 Storage Write Count: 10000
 Storage Write Throughput: 2160.72 writes/sec
 P50 Write Latency: 351.667µs
 P90 Write Latency: 423.166µs
 P99 Write Latency: 4.134375ms

 Index Latency Results:
 Indexing Throughput: 2144.91 events/sec
 P50 Index Latency: 0.042s
 P90 Index Latency: 0.054s
 P99 Index Latency: 0.065s
--- PASS: TestIntegrationBenchmarkResourceServer (4.87s)
```